service2/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
twilio==9.0.0
